requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
awscli==1.14.44
bleach==2.1.2
botocore==1.8.48
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
colorama==0.4.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
decorator==4.1.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
entrypoints==0.2.3.post1
graphviz==0.13
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.2
jsonpatch==1.24
jsonpointer==2.0
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
llvmlite==0.30.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.1.1
mistune==0.8.3
-e git+https://github.com/mattphillipsphd/ml_utils.git@0cec74c9388c30ae7acdc2fe747b1c322522b211#egg=ml_utils
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
notebook==5.2.2
numba==0.46.0
numpy==1.17.2
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.1.26
opt-einsum==3.0.1
pandas==0.25.1
pandocfilters==1.4.2
pbr==5.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.9.1
pyasn1==0.4.2
pycairo==1.16.2
pycocotools==2.0.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
pyro-ppl==0.4.1
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.12.0
sklearn==0.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboardX==1.8
terminado==0.7
testpath==0.3.1
torch==1.2.0
torchfile==0.1.0
torchvision==0.4.0
tornado==4.5.3
tqdm==4.35.0
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
umap-learn==0.3.10
unattended-upgrades==0.1
unidip==0.1.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
visdom==0.1.8.9
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
websocket-client==0.56.0
xkit==0.0.0
zope.interface==4.3.2
